fix: reposição grava custo+venda com margem; estoque por kg; code-splitting
- Reposição: marcar item comprado agora grava o preço PAGO em custoCompra e
  calcula a venda pela margem (mesma lógica da Nota IA). Antes sobrescrevia o
  preço de venda com o custo, zerando o lucro (mesmo bug da Nota IA).
- "Concluir lista" só repõe estoque (apenasEstoque), sem corromper preço/custo.
- ModalConfirmarCompra: campo de margem + venda editável + preview de lucro.
- Estoque por peso (kg) agora baixa na venda (converte gramas→kg); estoque
  e estoque_minimo migrados de int para numeric no schema.
- Módulo puro src/utils/estoque.js + 12 testes (estoque.test.mjs).
- Code-splitting das rotas (lazy/Suspense): bundle inicial 639→488 KB.
- Limpeza de lint: imports/vars mortos, catches comentados, remove api/ler-nota.js órfão.

+++ supabase/schema.sql
@@ -11,8 +11,8 @@ create table if not exists pdv_produtos (
   custo_compra float not null default 0,
   categoria text not null default 'mercearia',
   ativo boolean not null default true,
-  estoque int not null default 0,
-  estoque_minimo int not null default 0,
+  estoque numeric not null default 0,
+  estoque_minimo numeric not null default 0,
   codigos_alternativos jsonb default '[]'::jsonb
 );
 create index if not exists idx_pdv_produtos_loja on pdv_produtos(loja_id);
